requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
altair==4.2.2
anyio==3.6.2
apache-airflow==2.5.1
apache-airflow-providers-common-sql==1.3.3
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==2.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
blinker==1.5
boto3==1.26.65
botocore==1.29.65
cachelib==0.9.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.8
cryptography==39.0.1
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.19
ecdsa==0.18.0
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.92.0
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.4
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pathspec==0.9.0
pendulum==2.1.2
Pillow==9.4.0
plotly==5.13.0
pluggy==1.0.0
prison==0.2.1
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.3
pytest==7.2.1
python-daemon==2.3.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
python-nvd3==0.15.0
python-slugify==8.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
s3transfer==0.6.0
semver==2.13.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
starlette==0.25.0
streamlit==1.17.0
tabulate==0.9.0
tenacity==8.2.1
termcolor==2.2.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
yarl==1.8.2
zipp==3.12.1
bcrypt==4.0.1